requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
boto3==1.5.4
botocore==1.8.18
cachetools==2.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.4
dill==0.2.7.1
Django==2.0
django-js-asset==0.1.1
django-mptt==0.9.0
django-storages==1.6.5
docutils==0.14
Fabric==1.14.0
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==0.1.2
google-api-python-client==1.6.4
google-auth==1.2.1
google-cloud==0.31.0
google-cloud-bigquery==0.28.0
google-cloud-bigtable==0.28.1
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.0
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.0
google-cloud-pubsub==0.29.4
google-cloud-resource-manager==0.28.0
google-cloud-runtimeconfig==0.28.0
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.16.0
google-cloud-translate==1.3.0
google-cloud-videointelligence==1.0.0
google-cloud-vision==0.28.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.8.1
httplib2==0.10.3
idna==2.6
jmespath==0.9.3
mysqlclient==1.3.12
oauth2client==4.1.2
olefile==0.44
paramiko==2.4.0
Pillow==4.3.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.0.post1
psutil==5.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyNaCl==1.2.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
six==1.11.0
uritemplate==3.0.0
urllib3==1.22